Object file test support (#90)
* merge

* adding inline.o test

* tweaks

* tweaks

* getting things testing again

* debugging a test in GHA

* enable LFS pull in GitHub Actions

+++ include/orc/dwarf_structs.hpp
@@ -81,7 +81,7 @@ struct attribute_value {
     // If this attribute value has _both_, it is assumed
     // they are equal.
     int number() const {
-        return has(type::sint) ? sint() : uint();
+        return has(type::sint) ? static_cast<int>(sint()) : static_cast<int>(uint());
     }
     
     void string(pool_string x) {
